requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.1.0
Flask-Bcrypt==1.0.1